fix datatype for 64 bit architectures (cosmetic for now)
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@41482 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/httpd/httpd.h
@@ -122,7 +122,7 @@ struct mime_handler {
 	char *pattern;
 	char *mime_type;
 	char *extra_header;
-	void (*input) (char *path, webs_t stream, int len, char *boundary);
+	void (*input) (char *path, webs_t stream, unsigned long len, char *boundary);
 	void (*output) (unsigned char method, struct mime_handler * handler, char *path, webs_t stream);
 	int (*auth) (webs_t wp, int (*auth_check) (webs_t conn_fp));
 	unsigned char send_headers;
